refactor(llm): address reviewer feedback for twelvelabs integration
